Started work on db image url retreval
