+++ components/elements/paginationButtons.tsx
@@ -1,0 +1,9 @@
+function PaginationButtons(text: string) {
+    return (
+      <>
+        <button className="bg-black text-white"> {text} </button>
+      </>
+    );
+  };
+
+export default PaginationButtons;
