+++ components/types/images.ts
@@ -1,0 +1,3 @@
+export type Images = {
+    imageUris?: string[]
+  }
