+++ package.json
@@ -9,6 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "aws-sdk": "^2.1066.0",
     "next": "11.1.2",
     "next-share": "^0.12.2",
     "react": "17.0.2",
